requirements.txt
amqp==2.6.1
asgiref==3.2.10
Babel==2.8.0
billiard==3.6.3.0
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cssselect2==0.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1
django-allauth==0.42.0
django-debug-toolbar==2.2
django-heroku==0.3.1
gunicorn==20.0.4
html5lib==1.1
idna==2.10
importlib-metadata==1.7.0
kombu==4.6.11
numpy==1.19.1
oauthlib==3.1.0
Pillow==7.2.0
psycopg2==2.7.5
pycparser==2.20
Pyphen==0.9.5
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
tinycss2==1.0.2
tornado==5.1.1
urllib3==1.25.10
vine==1.3.0
webencodings==0.5.1
whitenoise==5.2.0
zipp==3.1.0
